+++ pynfact/dataman.py
@@ -7,6 +7,9 @@ Data and data structures management functions.
 
 .. versionadded:: 1.3.1a3
     Separate funcionality across multiple modules.
+
+.. versionchanged:: 1.3.1b1
+    Remove funcion ``and_array_in``: it has no use for this project.
 """
 
 
@@ -26,33 +29,12 @@ def or_array_in(dictionary, *keys):
     for key in keys:
         if key in dictionary:
             return dictionary[key]
-
     return None
-
-
-def and_array_in(dictionary, *keys):
-    """Check if every key in the array is in the dictionary.
-
-    The funcionality is the same as the keyword ``in``, but testing each
-    and every key in the array as an ``and`` operation.
-
-    :param keys: List of keys to test against the dictionary
-    :type keys: args
-    :param dictionary: Dictionary
-    :type dictionary: dict
-    :return: ``True`` if  every key is on the dictionary
-    :rtype: bool
-    """
-    for key in keys:
-        if key not in dictionary:
-            return False
-
-    return True
 
 
 def keys_exists(dictionary, *keys):
     """Check if nested keys exist in a dictionary.
-    
+
     :param dictionary: Nested dicionary
     :type dictionary: dict
     :param keys: Keys to test agains the dictionary