+++ pynfact/parsers/__init__.py
@@ -26,9 +26,8 @@
 # SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 ########################################################################
 """
-A blog-oriented static web content generator from Markdown to HTML5.
+PynFact! Parsers.
 
-:author: \J. A. Corbal
 :copyright: © 2012-2020, J. A. Corbal
 :license: MIT
 """
@@ -38,6 +37,8 @@ __email__ = "jacorbal@gmail.com"
 __license__ = "MIT"
 __maintainer__ = "J. A. Corbal"
 __status__ = "Production"
-__version_info__ = ('1', '3', '1b1')
-__version__ = '.'.join(__version_info__)
+
+
+from pynfact.parsers.parserrst import ParserRst
+from pynfact.parsers.parsermd import ParserMd