+++ pynfact/parsers/parsermd.py
@@ -8,33 +8,38 @@ Markdown translation to HTML and metadata retrieval.
 import markdown
 
 
-class Mulang:
-    """Generate HTML body from MarkUp LANGuage (mulang) source.
+class ParserMd:
+    """Generate HTML body from a Markdown source file.
 
     .. versionchanged:: 1.2.0a1
         Implement ``logging`` instead of printing to ``stdout`` and/or
         ``stderr``.
+
+    .. versionchanged:: 1.3.1a4
+        This class, formerly ``Mulang``, now it's just a parser for
+        Markdown since reStrucutedText support was added.
+
+    .. seealso:: :class:`ParserRst` and :mod:`parser`.
     """
 
-    def __init__(self, input_data, encoding='utf-8',
-                 output_format='html5', logger=None):
+    def __init__(self, input_data, encoding='utf-8', logger=None):
         """Constructor.
 
         This class has methods to parse the markup language and get the
         information.  The logging activity is only set to debug, since
-        it's not neccessary to tell continously the default activity.
+        it's not necessary to tell continuously the default activity.
 
         :param input_data: File from where the data is taken
         :type input_data: str
         :param encoding: Encoding the input file is in
         :type encoding: str
-        :param output_format: Output HTML version
-        :type output: str
         :param logger: Logger where to store activity in
         :type logger: logging.Logger
         """
         self.input_data = input_data
         self.encoding = encoding
+        self.logger = logger
+
         self.md = markdown.Markdown(
             extensions=['markdown.extensions.extra',
                         'markdown.extensions.toc',
@@ -44,34 +49,39 @@ class Mulang:
                         'markdown.extensions.codehilite',
                         'markdown.extensions.meta'],
             encoding=encoding,
-            output_format=output_format)
-        self.logger = logger
+            output_format='html5')
 
     def html(self):
-        """Generate HTML from Markdown data."""
-        input_file = open(self.input_data, mode="r",
-                          encoding=self.encoding)
-        text = input_file.read()
-        input_file.close()
-        html = self.md.convert(text)
+        """Generate HTML from a Markdown file."""
+        with open(self.input_data, "r", encoding=self.encoding) as f:
+            text = f.read()
 
+        html = self.md.convert(text)
         self.logger and self.logger.debug(
             'Parsed body data of: "{}"'.format(self.input_data))
 
         return html
 
     def metadata(self):
-        """Fetch metadata in the markdown file.
+        """Fetch metadata in a Markdown file.
+
+        The metadata syntax is case insensitive and as follows::
+
+            author: Author A. Author
+            title: This is a title
+            subtitle: A longer, much longer title for this entry
+            cdate: 2020-03-10
+            mdate: 2020-03-19
+            tags: this, are, a, lot, of, tags
+            comments: yes
 
         .. todo::
             Get the meta without generating HTML again.
         """
-        input_file = open(self.input_data, mode="r",
-                          encoding=self.encoding)
-        text = input_file.read()
-        input_file.close()
-        html = self.md.convert(text)
+        with open(self.input_data, "r", encoding=self.encoding) as f:
+            text = f.read()
 
+        html = self.md.convert(text)
         self.logger and self.logger.debug(
             'Parsed meta data of: "{}"'.format(self.input_data))